Refactor GitHubCrawlerTests
Split test class into unit and integration tests
